+++ mbrl/util/__init__.py
@@ -3,4 +3,9 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 from .logger import Logger
-from .replay_buffer import ReplayBuffer, SequenceTransitionIterator, TransitionIterator
+from .replay_buffer import (
+    ReplayBuffer,
+    SequenceTransitionIterator,
+    SequenceTransitionSampler,
+    TransitionIterator,
+)